requirements.txt
cython==0.29
networkx==2.5
numpy==1.19.2
matplotlib===3.3.4
seaborn==0.11.1
python-snappy==0.6.1
torch-scatter==2.0.5
torch-sparse==0.6.8
torch-geometric==1.7.0

